requirements.txt
google-genai
google-generativeai
dotenv
chainlit
